requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.0,<4.0
djangorestframework-authtoken>=1.0.0,<2.0
pandas>=1.1.0,<2.0
nltk>=3.5,<4.0